+++ CHANGELOG.md
@@ -11,6 +11,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   - Migrate unit tests to Pester 5
   - IPv6 preparation work
   - Increased code coverage to >85% fixes [#60](https://github.com/dsccommunity/DhcpServerDsc/issues/60).
+- `DhcpServerDsc.Common`
+  - Convert to buildable module with per file functions.
+- `DhcpServerDsc.OptionValueHelper`
+  - Convert to buildable module with per file functions.
 
 ### Added